requirements.txt
python-dotenv
pydantic
ollama
phidata
rich
requests
websockets
